# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPay API",
        "description": "Documentation de l'API OMPay (Orange Money Pay)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        },
        {
            "url": "https://ompay.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
        "/compte": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Afficher les informations du compte connecté",
                "operationId": "d25035bedc65bedc6212d189de192f12",
                "responses": {
                    "200": {
                        "description": "Informations du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Fall"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Awa"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "CPT-FAL12345"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "example": 50000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compte"
                ],
                "summary": "Créer un compte (client ou distributeur)",
                "description": "Permet de créer un compte utilisateur avec un numéro de compte unique.",
                "operationId": "fa582ec93e40adb9cd26524cc530d65b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "telephone",
                                    "password",
                                    "type"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Fall"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Awa"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "+221771234567"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "client",
                                            "distributeur"
                                        ],
                                        "example": "client"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Fall"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Awa"
                                                },
                                                "telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "CPT-FAL12345"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "example": 0
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/compte/solde": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Afficher le solde du compte connecté",
                "operationId": "53e6946c2ae2276b7f1e9b11d256d028",
                "responses": {
                    "200": {
                        "description": "Solde du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "example": 50000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/compte/history": {
            "get": {
                "tags": [
                    "Compte"
                ],
                "summary": "Lister les transactions du compte connecté",
                "operationId": "c21e7234d29d887e3df70903ed0a9324",
                "responses": {
                    "200": {
                        "description": "Historique des transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "transfert"
                                                    },
                                                    "montant": {
                                                        "type": "number",
                                                        "example": 1500
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "terminé"
                                                    },
                                                    "date_transaction": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-11-10T14:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}